requirements_w_appj_testbed.txt
ax-platform==0.2.5.1
botorch==0.6.4
crcmod==1.7
gpytorch==1.6.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
matlabengineforpython===R2021a
multipledispatch==0.6.0
numpy==1.22.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
pandas==1.4.1
pkg_resources==0.0.0
plotly==5.6.0
pyro-api==0.1.2
pyro-ppl==1.8.0
pyserial==3.5
python-dateutil==2.8.2
python-usbtmc==0.8
pytz==2021.3
pyusb==1.2.1
PyVISA==1.12.0
scikit-learn==1.0.2
scipy==1.8.0
seabreeze==2.0.3
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
torch==1.10.2
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.0.1
